Move cli/commands includes to the src
