+++ src/cli/commands/abstract_command.cpp
@@ -5,9 +5,9 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
-#include <utility>
+#include "abstract_command.hpp"
 
-#include <zephir/cli/commands/abstract_command.hpp>
+#include <utility>
 
 zephir::cli::commands::AbstractCommand::AbstractCommand(std::string name)
     : name_(std::move(name)) {}
