+++ src/cli/commands/abstract_command.hpp
