+++ src/cli/commands/api.cpp
@@ -5,9 +5,9 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
-#include <utility>
+#include "api.hpp"
 
-#include <zephir/cli/commands/api.hpp>
+#include <utility>
 
 zephir::cli::commands::ApiCommand::ApiCommand(std::string name)
     : AbstractCommand(std::move(name)),
