+++ src/cli/commands/build.cpp
@@ -5,9 +5,9 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
-#include <utility>
+#include "build.hpp"
 
-#include <zephir/cli/commands/build.hpp>
+#include <utility>
 
 zephir::cli::commands::BuildCommand::BuildCommand(std::string name)
     : AbstractCommand(std::move(name)),
