+++ src/cli/commands/fullclean.cpp
@@ -5,9 +5,9 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
-#include <utility>
+#include "fullclean.hpp"
 
-#include <zephir/cli/commands/fullclean.hpp>
+#include <utility>
 
 zephir::cli::commands::FullCleanCommand::FullCleanCommand(std::string name)
     : AbstractCommand(std::move(name)) {}
