+++ src/cli/commands/stubs.cpp
@@ -5,9 +5,9 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
-#include <utility>
+#include "stubs.hpp"
 
-#include <zephir/cli/commands/stubs.hpp>
+#include <utility>
 
 zephir::cli::commands::StubsCommand::StubsCommand(std::string name)
     : AbstractCommand(std::move(name)),
