+++ src/cli/commands/stubs.hpp
@@ -8,7 +8,7 @@
 #ifndef ZEPHIR_CLI_COMMANDS_STUBS_HPP_
 #define ZEPHIR_CLI_COMMANDS_STUBS_HPP_
 
-#include <zephir/cli/commands/abstract_command.hpp>
+#include "abstract_command.hpp"
 
 namespace zephir::cli::commands {
 struct StubsOptions {
