+++ src/cli/formatter.cpp
@@ -5,8 +5,9 @@
 // For the full copyright and license information, please view
 // the LICENSE file that was distributed with this source code.
 
+#include "formatter.hpp"
+
 #include <CLI/CLI.hpp>
-#include <zephir/cli/formatter.hpp>
 
 zephir::cli::Formatter::Formatter() = default;
