+++ tests/commands/build.options.cpp
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 
-#include <zephir/cli/application.hpp>
-#include <zephir/cli/commands/build.hpp>
+#include "cli/application.hpp"
+#include "cli/commands/build.hpp"
 
 using namespace zephir::cli::commands;
