+++ tests/commands/fullclean.options.cpp
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 
-#include <zephir/cli/application.hpp>
-#include <zephir/cli/commands/fullclean.hpp>
+#include "cli/application.hpp"
+#include "cli/commands/fullclean.hpp"
 
 using namespace zephir::cli::commands;
