+++ tests/commands/generate.options.cpp
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 
-#include <zephir/cli/application.hpp>
-#include <zephir/cli/commands/generate.hpp>
+#include "cli/application.hpp"
+#include "cli/commands/generate.hpp"
 
 using namespace zephir::cli::commands;
