+++ tests/commands/init.options.cpp
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 
-#include <zephir/cli/application.hpp>
-#include <zephir/cli/commands/init.hpp>
+#include "cli/application.hpp"
+#include "cli/commands/init.hpp"
 
 using namespace zephir::cli::commands;
