+++ tests/commands/stubs.options.cpp
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 
-#include <zephir/cli/application.hpp>
-#include <zephir/cli/commands/stubs.hpp>
+#include "cli/application.hpp"
+#include "cli/commands/stubs.hpp"
 
 using namespace zephir::cli::commands;
